FIX - Character position "freeze" / Adjust player speed when entering/exiting buildingsduring movement with simultaneous use of movement keys + R/Shift
POK12:
Description : We have identified a bug in our game that causes a character's position to freeze during movement when the player simultaneously presses movement keys and the R/Shift key. This not only creates a freeze for other players on the character's sprite but also causes a visible position shift. This bug is affecting the gameplay and needs to be fixed as soon as possible.
POK6 :
Currently, when the player enters or exits a building, their movement speed remains the same regardless of their current speed setting. This can lead to inconsistency and may cause issues when players try to navigate around indoors. To address this, we propose adding a check for the "canCrossRun" setting when the player enters/exits a building. If "canCrossRun" is true, the player's speed should remain unchanged. Otherwise, their speed should be reset to the default walking speed of 3.

In addition, we propose adding a constant for the player speed in the "GameInfos" file. This constant should have three values: the default walking speed of 3, the sprinting speed of 5, and the "canCrossRun" speed of 7. This will allow for more flexibility in adjusting the player's speed and simplify the implementation of this feature.